+++ src/ethereum/EthereumEngine.ts
@@ -1,5 +1,5 @@
 import { Common } from '@ethereumjs/common'
-import { TransactionFactory } from '@ethereumjs/tx'
+import { TransactionFactory, TypedTxData } from '@ethereumjs/tx'
 import { add, ceil, div, gt, lt, lte, mul, sub } from 'biggystring'
 import { asMaybe, asObject, asOptional, asString } from 'cleaners'
 import {
@@ -1218,12 +1218,12 @@ export class EthereumEngine extends CurrencyEngine<
       this.networkInfo.supportsEIP1559 === true
         ? {
             maxFeePerGas: otherParams.gasPrice,
-            maxPriorityPerGas: otherParams.minerTip
+            maxPriorityFeePerGas: otherParams.minerTip
           }
         : { gasPrice: otherParams.gasPrice }
 
     // Transaction Parameters
-    const txParams = {
+    const txParams: TypedTxData = {
       nonce: nonceHex,
       ...txFeeParams,
       gasLimit: gasLimitHex,