+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
+- fixed: Incorrect spelling for EIP-1559 parameter (maxPriorityFeePerGas)
+
 ## 2.18.2 (2023-12-18)
 
--fixed: Various regressions and bugs caused by recent refactors
+- fixed: Various regressions and bugs caused by recent refactors
 
 ## 2.18.1 (2023-12-18)