+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.18.3 (2023-12-20)
+
 - added: `minerTip` to `feeRateUsed` for EIP-1559 transaction
 - changed: Added back EIP1559 support to Fantom
 - fixed: Fallback query for baseFee if not cached for EIP-1559 currencies